requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
pymysql==1.1.0
jinja2==3.1.2